requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
supabase==2.0.2
psycopg2-binary==2.9.7
pandas==2.3.1
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
reportlab==4.0.4
Pillow==10.2.0
Werkzeug==2.3.7
gunicorn==21.2.0
lxml==4.9.3
chardet==5.2.0
urllib3==2.0.7
certifi==2023.7.22
idna==3.4
charset-normalizer==3.3.0
six==1.16.0
pypdf==3.5.1
pdfplumber==0.9.0
markdown==3.5.1
jinja2==3.1.2
markupsafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
python-multipart==0.0.6
numpy==2.3.2
huggingface_hub==0.20.3
html5lib==1.1
openai==1.3.8
serpapi==0.1.5
flask-compress==1.13
chardet==5.2.0
redis==4.5.4
flask-socketio==5.3.0

beautifulsoup4
html5lib
lxml
newspaper3k
readability-lxml
trafilatura
weasyprint
newspaper3k
readability-lxml
trafilatura
beautifulsoup4
flask
newspaper3k
readability-lxml
requests
trafilatura
beautifulsoup4
blinker
certifi
chardet
charset-normalizer
click
flask
flask-compress
flask-cors
flask-socketio
google-generativeai
gunicorn
html5lib
huggingface-hub
idna
itsdangerous
jinja2
lxml
markdown
markupsafe
newspaper3k
numpy
openai
openpyxl
pandas
pdfplumber
Pillow
pypdf
PyPDF2
python-docx
python-dotenv
python-multipart
readability-lxml
redis
reportlab
requests
serpapi
six
supabase
trafilatura
urllib3
Werkzeug
beautifulsoup4
blinker
certifi
chardet
charset-normalizer
click
flask
flask-compress
flask-cors
flask-socketio
google-generativeai
gunicorn
html5lib
huggingface-hub
idna
itsdangerous
jinja2
lxml
markdown
markupsafe
newspaper3k
numpy
openai
openpyxl
pandas
pdfplumber
Pillow
pypdf
PyPDF2
python-docx
python-dotenv
python-multipart
readability-lxml
redis
reportlab
requests
serpapi
six
supabase
trafilatura
urllib3
Werkzeug
